Added precise key events

+++ project/src/ExternalInterface.cpp
@@ -2663,14 +2663,14 @@ namespace lime {
 	}
 
 
-	long lime_sdl_get_ticks () {
+	int lime_sdl_get_ticks () {
 
 		return System::GetTicks ();
 
 	}
 
 
-	HL_PRIM long HL_NAME(hl_sdl_get_ticks) () {
+	HL_PRIM int HL_NAME(hl_sdl_get_ticks) () {
 
 		return System::GetTicks ();
